Support use of i18n middleware
In order to properly support multi-language apps the translate method needs to be read from request, as set by i18n middleware, so it can read the `Accept` language headers from the request to determine the target langauge.

By passing in a static method to these middlewares the only language available is the default fallback language - `en` - and so multi-language support is not possible.

By allowing `req.translate` to be used to perform translations we can then allow non-en languages to be used.